feat(devkit): add mapRemotes helper
