+++ packages/devkit/src/utils/module-federation/index.ts
@@ -1,4 +1,5 @@
 export * from './share';
 export * from './dependencies';
 export * from './package-json';
+export * from './remotes';
 export * from './models';
